Added implementation and initial specs
The final goal is to rely solely on active_model/lint.rb
to perform the compliance tests. However, lint.rb still
needs to be updated to cover the intended behavior more
precisely.

Nevertheless, next step is to include active_model as a
development dependency and run both lint.rb and our own
specs for the time being.